python/mongo-parquet/pyproject.toml
[project]
name = "mongo_parquet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.1.3",
    "polars>=1.14.0",
    "pydantic>=2.10.6",
    "pymongo>=4.10.1",
    "pymongoarrow>=1.6.0",
    "python-dotenv>=1.0.1",
    "zstandard>=0.23.0",
    "adlfs>=2024.12.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"